Rename e2e util files
